Increase Render Distance
- Increase base draw distance to 5000.
- Increase max draw distance to 15000.
- Add medium draw distance for certain levels.
- Remove cyclorama from depth buffer.
- Remove cyclorama requirement from gLevelClearColor.
- Change the type of gNumEnemyOfKind to int32_t.
- Raise MAX_SUPERTILE_ACTIVE_RANGE to 20.
- Change all values in gLevelSuperTileActiveRange to fit in the 8-20 range.
- Raise MAX_LIQUID_MESHES to 255.
- Use proper copyright symbol in Windows resource file.
- Update documentation.
Issue https://github.com/jorio/Bugdom/issues/24.

+++ src/Enemies/Enemy.c
@@ -29,7 +29,7 @@ static Boolean KillEnemy(ObjNode *theEnemy);
 /*    VARIABLES      */
 /*********************/
 
-short		gNumEnemyOfKind[NUM_ENEMY_KINDS];
+int32_t		gNumEnemyOfKind[NUM_ENEMY_KINDS];
 short		gNumEnemies;
